refactor: MyProjectPosts 중 '참여프로젝트' 영역 분리
